Add AnnotateModels::SchemaInfo::IndexInfo
